Revert "Bug 2033073 - Retry connections that fail with SSL_ERROR_DECRYPT_ERROR_ALERT  a=dmeehan" also reverting Bug 2024146
This reverts commit bb61e518e1334953771cff15791fd53f02a243f9.
